Fix problems on ELF systems

+++ runtime/debug.h
@@ -55,7 +55,7 @@ extern CHEETAH_INTERNAL unsigned int debug_level;
 // #define CILK_STACKFRAME_MAGIC 0xCAFEBABE
 
 CHEETAH_INTERNAL void set_alert_level_from_str(const char *const);
-CHEETAH_INTERNAL void set_alert_level(unsigned long);
+CHEETAH_INTERNAL void set_alert_level(unsigned int);
 CHEETAH_INTERNAL void set_debug_level(unsigned int);
 CHEETAH_INTERNAL void flush_alert_log(void);
